Fix: lower manifest version back down to 2
Lowers the manifest version back from 3 to 2,
as for some reason the manifest version was
incremented to 3 without actually updating
the manifest to support version 3. This is just
a temporary fix as Google is looking to
deprecate MV2 sometime after MV3 is stable.
Additionally MV3 is still in preview/alpha and
it is not currently needed to update.