Add: support for remote streaming
Follows my theory in PikaDude/ac-music-extension-revived#5
Songs are now stored remotely on my server, and
are cached locally on the client's computer
for 5 days. They will continue to work offline even
if the song changes, or the user restarts chrome.
The only catch is that the user must listen to the
song at least once for it to be cached, otherwise
it will not play. Additionally, the songs must be
renewed every 5 days, so the offline cache is not
permanent. 5 days was just a test duration, I can
extend it if we deem it necessary.